refact: Move FieldOption class into Form namespace
