--- tests/Field/TestCase.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Kirby\Field;
-
-use Kirby\TestCase as BaseTestCase;
-
-class TestCase extends BaseTestCase
-{
-}
